event_service/requirements.txt
fastapi[standard]
uvicorn
psycopg2-binary
python-dotenv